refact: add dto package and move CarDTO
